adds hiera support and keeps from clobbering hiera.yaml
This commit builds on Branan's commit that adds hiera support and cleans
up the rake task so that you can modify the hiera.yaml without it being
rewritten on each run.

Also adds %{test_suite} to the hiera hierarchy, which allows you to specify
a fact, test_suite, and have hiera look in a specific YAML file.