(Maint) Rename PuppetlabsSpec::Puppet{Seams,Internals}
We've added more methods to obtain instances of internal Puppet objects.
These additions make the name PuppetSeams more awkward than it already
was.

This patch fixes the problem of the awkward name by renaming the module
to `PuppetlabsSpec::PuppetInternals`  This module is to be used in
tests to provide easier (Puppet version agnostic) access to the
scope, resource, node, and compiler method.

Paired-with: Andy Parker <andy@puppetlabs.com>