(#15464) Add gemspec from 0.3.0 published gem
Without this patch the puppetlabs_spec_helper cannot be used as a Gem
library easily.  This is a problem because we need this repository
usable from bundler and from Git source when doing local development and
testing.

This gemspec is actually incorrect, it should use the same versions of
rspec as does Puppet and Facter.