(#15464) Make contributing easy via bundle Gemfile
Without this patch the process of figuring out how to quickly set up a
development and testing environment.  This patch addresses the problem
by providing a Bundler compatible Gemfile that specifies all of the Gem
dependencies.

Contributors can now easily get a working development and testing
environment using this sequence of commands:

    $ git clone git://github.com/puppetlabs/puppetlabs_spec_helper
    $ cd puppetlabs_spec_helper
    $ bundle exec rspec

The .noexec.yaml file excludes the `rake` command so that the Gemfile
doesn't raise an exception if the `rubygems-bundler` Gem is installed and
automatically running `bundle exec` for us.