Update pinned version for gemspec
This commit updates the pinned version for the
gemspec dependencies to match the dependencies
from the Gemfile.

The issue is that the version was not restrictive enough
for the gemspec. This means that it started pulling in
rspec 3 as a dependency as soon as it was released.

Since many projects rely on this gem to standarize library
versions, it needs to be updated to be pinned properly to
a 2.x version.