Add mocha mocking back
One of puppetlabs_spec_helpers "features" is to enable mocha mocking
support. The dependency on mocha was removed in 0.5.0 and references to
it were removed, but this caused changes in the testing behavior.

Not that mocha is the recommended mocking library (and this "feature"
should probably be removed in the future anyway), but this commit adds
it back to undo the breaking change.