Fix rspec 3.0 error when loading spec_helper
Without this patch the spec helper fails to load with RSpec 3.0 because
a method that does not exist is trying to be aliased.  This is a problem
because the spec helper is incompatible with RSpec 3.0 and it would be
nice to support the latest rspec version when developing modules.

This patch addresses the problem by aliasing the methods only if they
exist.