(MODULES-2002) Accept alternate fixtures file
Without this patch, the location of the fixtures YAML file is hardcoded
to be ".fixtures.yaml."  This does not accomodate modules that are
tested across environments with differing module locations.

This patch introduces an environment variable FIXTURES_YML, which can
be used to specify the location of an alternate fixtures data file.