Support loading from .fixtures.yaml too
YAML can have multiple extensions, .yml and .yaml. Hiera has used
.yaml all over the place but for some reason the spec_helper gets
extremely sad when you have a fixtures.yaml instead. This is
annoying for the end user and easily dealt with in code.

Additionally we now actually catch Psych::SyntaxError and abort in
that case instead of just checking for `Error::ENOENT` and returning
before. I'm not even sure if the `if not fixtures` would ever be
triggered since a syntax error would actually raise an error we
were not rescuing.