Ability to split spec files across multiple CI nodes
With many specs it can speed up tests considerably to use a CI matrix
to run tests in parallel.  This adds CI_NODE_TOTAL and CI_NODE_INDEX
environment variables to the rake spec_standalone task to run a subset
of the spec files on a single node in a simple manner.

The index variable should be added to the CI matrix, e.g. a Travis CI
"env" section or a Jenkins matrix job.