(maint) Add task to print development version
This adds the task `compute_dev_version` to print a computed
development version of the module.

The development version is computed by appending the short sha of the
repo to the current version specified by the module. If the build is
ran in a CI environment (one with the variables BUILD_NUMBER or
TRAVIS_BUILD_NUMBER set) it will prepend the build number to the short
sha so that artifacts can easily be sorted by build numbers.

Note the printing of this dev version does not include a trailing
newline. The proposed workflow of it would be:
```
BLACKSMITH_FULL_VERSION=`bundle exec rake compute_dev_version`
export BLACKSMITH_FULL_VERSION
bundle exec rake module:bump:full
bundle exec rake module:push
```