Remove RuboCop to fix spec testing on Ruby 1.8
Ruby 1.8 is deprecated, but is still supported by PE 3.8 which is still maintained.
As @alexjfisher pointed out in the discussion following up the merge of PR #140,
RuboCop (or at least one of its dependencies) is not compatible with Ruby 1.8.
Hence spec testing on Ruby 1.8 will break because `bundle install` fails.

Since the GemSpec is generated with Jeweler which does not support different
gem dependencies for different Ruby versions, I suggest remove RuboCop for the moment.