Accurate 'spec' description
It bothers me that the description says `:spec` is run against a clean fixtures directory. It does not run `:spec_clean` first, so if your fixtures are even partially populated, the description is misleading.. Either the description should be updated, or `:spec_clean` should be run first and last. I do not believe that the behavior change would be desired by most users, though. Instead, the description has been updated for accuracy.