(MODULES-4394) Make the module_working_dir configurable
Previously the module_work_dir used when installing modules fixtures via the
forge used a hardcoded relative path inside of the module.  However during CI
testing this path gets very long and in some instances on Windows, the fixtures
fail to install.  This commit adds support for an environment variable called
`MODULE_WORKING_DIR` which can be set and will instruct the helper to use this
user set directory.  This commit preserves converting the path into a Puppet
compatible string due to PUP-4884.  If the environment variable is not set the
current default of `spec/fixtures/work-dir` is used.