(maint) Only output gettext-setup gem errors if rake is verbose
Previously loading the rake task would also generate text output if the
the gettext-setup was not loaded or could not be configured.  However this
breaks the compute-dev-version rake task as it should output a single line of
the module version.  This commit confines the helpful messages to only be output
if `--verbose` is specified on the Rake command line.  By default verbose is not
set.