(MODULES-4471) Add CI_SPEC_OPTIONS environment variable to modify rspec
Previously it was not possible to modify the behaviour of RSpec from the rake
tasks.  This commit adds an additional environment variable called
CI_SPEC_OPTIONS which will be appended to the rspec invocation so that additional
settings, such as, output formatting can be passed into the rake tasks.

This commit also adds documentation to the README about CI_SPEC_OPTIONS.