Add dependency on parallel_tests
The rake task release_checks uses the parallel_tests so this should be
either a dependency or the task should fall back on regular tests. Given
parallel_tests works well, I decided on the former.

Note that since parallel_tests depends on ruby >= 2.0 this also means
puppetlabs_spec_helper depends on that.