(PDK-1033) Use `--unshallow` when updating a shallow clone
git does not fetch beyond the original depth in shallow repositories. This change uses `--unshallow` when a `ref` is specified, to avoid the issues shown in PDK-1033, namely not being able to switch from a shallow non-`ref` fixture clone to one with a `ref` to outside the original clone.