(MODULES-7856) Allow optional repositories based on puppet version
Some types/providers maybe missing when testing against puppet6 as a
gem. Allow an optional repository to be specified which will only be
installed if the current puppet version matches the semantic version
constraint. For example, stdlib will always be installed, but
selinux_core is only installed when testing on puppet6 or above:

    repositories:
      stdlib: https://github.com/puppetlabs/puppetlabs-stdlib.git
      selinux_core:
        repo: https://github.com/puppetlabs/puppetlabs-selinux_core.git
        puppet_version: ">= 6.0.0"