(PDK-997) Remove Dir.chdir call from check:test_file task
This `chdir` breaks the isolation of the rake task and affects the
behaviour of other tasks that are invoked in it in the same `rake`
execution.

Rather than modify the task so that it changes back to the parent
directory after the task has finished, we can just remove the `chdir`
entirely and just change the glob to look for files in the directory.

This PR also adds some basic unit tests for the `check:test_file` task.