(PDK-1717) Add guard clause to module path dir enum loop
Prior to this commit, If a non-existent directory is specified in
the `MODULEPATH` env var, it will cause the enumeration of module
paths to fail.

This commit performs a `Dir.exist?` check in that loop before any
subsequent unguarded / unrescued calls are made.

This issue manifested due to a conflict with the Linux package
`lmod` also using the `MODULEPATH` env var and populating it with
non-existent dir paths.