(GH-372) Reset min Ruby requirement
Prior to this change the required Ruby version was set to greater than
or equal to 2.7. This meant that the spec_helpers gem would not run on
systems with Ruby 2.5.

This change updates the required Ruby version so that Rubies 2.5 and
above are compatibel.

Additionally the following gems were downgraded for compatibility:

* pathspec is now ~> 0.2
* puppet-lint is now ~> 2.5.2