(maint) Relax Ruby requirement
Prior to this commit, puppetlabs_spec_helper required Ruby >= 2.7 in its
Gemspec.

Puppetserver 7, which is expected to be supported as part of PE 2021.7.z
LTS through 2025, uses JRuby 9.3.y.z which targets compatibility with
Ruby 2.6 and uses that in its RUBY_VERSION.

Also, Puppet 7 lists its required Ruby as >= 2.5 in the Gemspec.

This commit relaxes the gemspec's Ruby requirements from >= 2.7 to >=
2.5 to allow full compatibility with Puppet and Puppetserver 7.