(GH-397) - Honour default symlink when additional symlinks delcared
Before, declaring a symlink would overwrite the symlink to the current
module.
This is not what we would want for testing, and we should be able to
declare additional symlinks to the default one.
Now, we extra symlinks declared, we merge into the auto_symlink, if not
operate as before.