(CAT-2113) Adjusting minimum ruby version required
Following investigation on an issue that prevented puppet-resource_api
from consuming latest puppetlabs_spec_helper, it was noted that this repository
assumes Ruby 2.7 as the minimum required version, causing some repositories to
fail their testing environment builds due to the fact that they are consuming
some form of Ruby 2.6 in order to test properly against puppetserver.

This commit adjusts the required Ruby version to the standard used across other
modules/tools of the ecosystem.